% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_bmi_icd}
\alias{assign_bmi_icd}
\title{assign_bmi_icd ----------------------------------------------------------}
\usage{
assign_bmi_icd(bmi)
}
\arguments{
\item{bmi}{body mass index}
}
\value{
vector of assigned icd bmi codes
}
\description{
assign_bmi_icd ----------------------------------------------------------
}
