% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_bmi_status}
\alias{assign_bmi_status}
\title{assign_bmi_status -------------------------------------------------------}
\usage{
assign_bmi_status(bmi)
}
\arguments{
\item{bmi}{body mass index}
}
\value{
vector of assigned bmi statuses
}
\description{
assign_bmi_status -------------------------------------------------------
}
