% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bmi_adult}
\alias{bmi_adult}
\title{Calculate Body Mass Index (BMI) for Adults}
\usage{
bmi_adult(weight, height)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}
}
\value{
vector of BMI values (kg/m2)
}
\description{
Calculate Body Mass Index (BMI) for Adults
}
\examples{
bmi_adult(70, 160)
}
