% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex_medline.R
\name{codex_medline}
\alias{codex_medline}
\title{Search the MedlinePlus Connect API}
\source{
U.S. National Library of Medicine
}
\usage{
codex_medline(
  code = NULL,
  code_system = NULL,
  code_title = NULL,
  language = "english",
  format = "json"
)
}
\arguments{
\item{code}{diagnosis (problem) code, medication code, laboratory test code, or procedure code}

\item{code_system}{ICD-9-CM, ICD-10-CM, CPT, SNOMED, RXCUI, or NDC}

\item{code_title}{You may also identify the name/title of the problem code. However, this information does not impact the response.}

\item{language}{Identify if you would like the response to be in English or Spanish. MedlinePlus Connect will assume English is the language if it is not specified.}

\item{format}{JSON, JSONP, or XML}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
MedlinePlus Connect is a free service of the National Library of Medicine (NLM), National Institutes of Health (NIH), and the Department of Health and Human Services (HHS). This service allows health organizations and health IT providers to link patient portals and electronic health record (EHR) systems to MedlinePlus, an authoritative up-to-date health information resource for patients, families, and health care providers.
}
\details{
TMedlinePlus Connect accepts and responds to requests for information based on diagnosis (problem) codes, medication codes, laboratory test codes, and medical procedure codes. When an EHR, patient portal or other system submits a code-based request, MedlinePlus Connect returns a response that includes links to patient education information relevant to the code.
}
\note{
\href{https://medlineplus.gov/medlineplus-connect/web-service/}{MedlinePlus Connect API}
}
\examples{
codex_medline(code_title = "Chantix 0.5 MG Oral Tablet", code_system = "rxcui")

codex_medline(code = "00310-0751-39", code_system = "ndc")
}
