% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codexchain-package.R
\docType{package}
\name{codexchain-package}
\alias{codexchain}
\alias{codexchain-package}
\title{codexchain: Access Public Medical Coding Data APIs}
\description{
The goal of codexchain is to provide access to a variety of public medical coding APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/codexchain}
  \item Report bugs at \url{https://github.com/andrewallenbruce/codexchain/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com} [copyright holder]

}
\keyword{internal}
