% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex_mue.R
\name{codex_mue}
\alias{codex_mue}
\title{Search CMS' NCCI Medically Unlikely Edits (MUEs) API}
\usage{
codex_mue()
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=codex_mue]{codex_mue()}} allows you to search CMS' NCCI
Medically Unlikely Edits (MUEs) API.
}
\section{NCCI Medically Unlikely Edits (MUEs)}{
Medically Unlikely Edits (MUEs) define for each HCPCS / CPT code the
maximum units of service (UOS) that a provider would report under most
circumstances for a single beneficiary on a single date of service.
Practitioner services also refers to ambulatory surgical centers. DME
refers to provider claims for durable medical equipment.

The CMS National Correct Coding Initiative (NCCI) promotes national
correct coding methodologies and reduces improper coding which may
result in inappropriate payments of Medicare Part B claims and Medicaid
claims. NCCI procedure-to-procedure (PTP) edits define pairs of
Healthcare Common Procedure Coding System (HCPCS)/Current Procedural
Terminology (CPT) codes that should not be reported together for a
variety of reasons.

The purpose of the PTP edits is to prevent improper payments when
incorrect code combinations are reported. The edits in this dataset
are active for the dates indicated within. This file should NOT be used
by state Medicaid programs as their edit file. For more information,
visit https://www.medicaid.gov/medicaid/program-integrity/ncci/index.html.
\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.medicaid.gov/dataset/552def06-89ab-5f6f-a6de-90b27a792127}{CMS' NCCI Medically Unlikely Edits (MUEs) API}
}
}
}

\examples{
\dontrun{
# Returns the first 10,000 results

codex_mue()
}
}
