% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex_ptp.R
\name{codex_ptp}
\alias{codex_ptp}
\title{Search CMS' NCCI Procedure to Procedure Edits (PTP) API}
\usage{
codex_ptp(explain = FALSE)
}
\arguments{
\item{explain}{If \code{TRUE}, converts data to a more human-readable form;
default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=codex_ptp]{codex_ptp()}} allows you to search CMS' NCCI Procedure
to Procedure Edits (PTP) API.
}
\section{NCCI Procedure to Procedure Edits (PTP)}{
The CMS National Correct Coding Initiative (NCCI) promotes national
correct coding methodologies and reduces improper coding which may result
in inappropriate payments of Medicare Part B claims and Medicaid claims.
NCCI procedure-to-procedure (PTP) edits define pairs of Healthcare Common
Procedure Coding System (HCPCS)/Current Procedural Terminology (CPT) codes
that should not be reported together for a variety of reasons. The purpose
of the PTP edits is to prevent improper payments when incorrect code
combinations are reported. Practitioner services also refers to ambulatory
surgical centers. DME refers to provider claims for durable medical
equipment.
\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.medicaid.gov/dataset/bf3c4970-7dce-49df-b555-bb62ccf9ffb6}{CMS' NCCI Procedure to Procedure Edits (PTP) API}
}
}
}

\examples{
\dontrun{
# Returns the first 10,000 results

codex_ptp()

codex_ptp(explain = TRUE)
}
}
