% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codexchain-package.R
\docType{package}
\name{codexchain-package}
\alias{codexchain}
\alias{codexchain-package}
\title{codexchain: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/codexchain}
  \item Report bugs at \url{https://github.com/andrewallenbruce/codexchain/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com}

}
\keyword{internal}
